fix(client): remove full-screen-loader from App and implement in WorkspaceLayout
