+++ packages/amplication-client/src/Workspaces/WorkspaceLayout.tsx
@@ -1,6 +1,6 @@
-import { CircularProgress } from "@amplication/design-system";
+import { AnimationType, FullScreenLoader } from "@amplication/design-system";
 import { StiggProvider } from "@stigg/react-sdk";
-import React, { lazy, useEffect, useState } from "react";
+import React, { lazy, useCallback, useEffect, useState } from "react";
 import { isMobileOnly } from "react-device-detect";
 import { match } from "react-router-dom";
 import { useTracking } from "react-tracking";
@@ -90,6 +90,13 @@ const WorkspaceLayout: React.FC<Props> = ({ innerRoutes, moduleClass }) => {
     loadingCreateMessageBroker,
   } = useResources(currentWorkspace, currentProject, addBlock, addEntity);
 
+  const [keepLoadingAnimation, setKeepLoadingAnimation] =
+    useState<boolean>(true);
+
+  const handleTimeout = useCallback(() => {
+    setKeepLoadingAnimation(false);
+  }, []);
+
   const { trackEvent, Track } = useTracking<{ [key: string]: any }>({
     workspaceId: currentWorkspace?.id,
     projectId: currentProject?.id,
@@ -120,7 +127,7 @@ const WorkspaceLayout: React.FC<Props> = ({ innerRoutes, moduleClass }) => {
     }
   }, [currentWorkspace]);
 
-  return currentWorkspace ? (
+  return currentWorkspace && !keepLoadingAnimation ? (
     <AppContextProvider
       newVal={{
         currentWorkspace,
@@ -201,7 +208,11 @@ const WorkspaceLayout: React.FC<Props> = ({ innerRoutes, moduleClass }) => {
       )}
     </AppContextProvider>
   ) : (
-    <CircularProgress centerToParent />
+    <FullScreenLoader
+      animationType={AnimationType.Full}
+      minimumLoadTimeMS={2000}
+      onTimeout={handleTimeout}
+    />
   );
 };
